Add unit testing in cart.spec.js
