+++ config/config.json
@@ -8,11 +8,12 @@
     "operatorsAliases": "Op"
   },
   "test": {
-    "username": "travis",
-    "database": "ec_web_test",
+    "username": "root",
+    "password": "12345678",
+    "database": "ec_web",
     "host": "127.0.0.1",
     "dialect": "mysql",
-    "logging": false
+    "operatorsAliases": "Op"
   },
   "production": {
     "use_env_variable": "JAWSDB_URL",
